% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioLake-package.R
\docType{package}
\name{PhysioLake-package}
\alias{PhysioLake}
\alias{PhysioLake-package}
\title{PhysioLake: versioned, lineage-tracked data management for the Physio ecosystem}
\description{
A thin integration layer over OmicsLake. See the package README and
\link{PhysioExperimentAdapter} / \link{physioPut}.
}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
